moved data to constants
