+++ src/components/Dashboard/dream/create/Characters.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { writable } from 'svelte/store';
 
-  import dreamData from '../../../../data/dream';
+  import dreamData from '@constants/dream';
   import { tablePrompt } from '@stores/dream';
 
   import Dropdown from './Dropdown.svelte';
